Split compat layer into different source files #806
