--- mz_compat_shim.h.in
@@ -1,12 +1,0 @@
-/* @FILE_H@ -- Compatibility layer shim
-   part of the minizip-ng project
-
-   This program is distributed under the terms of the same license as zlib.
-   See the accompanying LICENSE file for the full text of the license.
-*/
-#ifndef @MZ_COMPAT_FILE@
-#define @MZ_COMPAT_FILE@
-
-#include "mz_compat.h"
-
-#endif
